css/css-syntax/tokenizing-and-parsing/input-byte-stream/i18n-css-syntax-bytes-020.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Syntax, input byte stream: http vs. charset link</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-syntax/#input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http dom'>
<meta name="assert" content="An HTTP encoding declaration for a stylesheet takes precedence over a charset attribute in the link element.">
<style type='text/css'>
.test div { width:50px; }

p { color: gray; margin-top: 3em; font-size: 80%; }
</style>
<link rel='stylesheet' charset='utf-8' type='text/css' href='support/http1.css' /></head>
<body>



<div class="test"><div id="box" class="ýäè">&#xA0;</div></div>


<!--Notes:
This test is only valid if the user agent recognises the HTTP and charset link declarations.

The HTML link element contains a charset attribute that says that the style sheet is UTF-8, but the HTTP header for the style sheet says ISO 8859-1. If the HTTP header overrides the charset attribute, the selector will match the class name and the test will pass.

-->
<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>

<div id='log'></div>

<p class="ýäè"><span><span id="debug">Text as read from the style sheet: </span><br />(If blank, the style sheet hasn't been read at all.)</span></p>
</body>
</html>
